refactor: split build_mpt to separate file
